+++ frontend/include/hipdnn_frontend/backend/hipdnn_backend_interface.hpp
@@ -52,7 +52,7 @@ public:
     virtual void logging_callback_ext(hipdnnSeverity_t severity, const char* msg) = 0;
 
     // NOLINTBEGIN
-    virtual hipdnnStatus_t set_plugin_paths_ext(int64_t num_paths,
+    virtual hipdnnStatus_t set_plugin_paths_ext(size_t num_paths,
                                                 const char* const plugin_paths[],
                                                 hipdnnPluginLoadingMode_t mode)
         = 0;
